requirements.txt
# Core ML Libraries
accelerate==0.33.0
datasets
diffusers==0.33.1     
transformers==4.44.2
peft==0.15.2

# Essential Dependencies
bitsandbytes==0.45.3
safetensors>=0.4.0,<0.6.0 
Pillow>=9.0.0
numpy<2.0.0,>=1.22.0    
fsspec<=2024.12.0,>=2023.1.0
packaging<=24.2        
six<=1.16           
tqdm>=4.60.0
datasets>=3.0.0  

# Optional/Tooling
gradio>=4.0.0
wandb
tensorboard
cd

# Docker
jupyterlab
gdown
wget